#129 test: 토큰 프로바이더 테스트 추가 및 패키지 이동
